Align cncf provider file names with AIP-21 Changes to import paths:
`operators.kubernetes_pod` -> `operators.pod`
`triggers.kubernetes_pod` -> `triggers.pod`

This PR is backward compatible
