+++ scripts/in_container/verify_providers.py
@@ -148,6 +148,7 @@ KNOWN_COMMON_DEPRECATED_MESSAGES: set[str] = {
     "Deprecated API features detected! These feature(s) are not compatible with SQLAlchemy 2.0",
     "Implementing implicit namespace packages (as specified in PEP 420) is "
     "preferred to `pkg_resources.declare_namespace`",
+    "This module is deprecated. Please use `airflow.providers.cncf.kubernetes.operators.pod` instead.",
 }
 
 # The set of warning messages generated by direct importing of some deprecated modules. We should only
@@ -161,6 +162,8 @@ KNOWN_DEPRECATED_DIRECT_IMPORTS: set[str] = {
     "numpy.ufunc size changed, may indicate binary incompatibility.",
     "This module is deprecated. Please use `airflow.providers.amazon.aws.operators.lambda_function`.",
     "S3ToSnowflakeOperator is deprecated.",
+    "This module is deprecated. Please use `airflow.providers.cncf.kubernetes.operators.pod` instead.",
+    "This module is deprecated. Please use `airflow.providers.cncf.kubernetes.triggers.pod` instead.",
 }
