+++ tests/kubernetes/test_kubernetes_helper_functions.py
@@ -23,7 +23,7 @@ import pytest
 from pytest import param
 
 from airflow.kubernetes.kubernetes_helper_functions import create_pod_id
-from airflow.providers.cncf.kubernetes.operators.kubernetes_pod import _create_pod_id
+from airflow.providers.cncf.kubernetes.operators.pod import _create_pod_id
 
 pod_name_regex = r"^[a-z0-9]([-a-z0-9]*[a-z0-9])?(\.[a-z0-9]([-a-z0-9]*[a-z0-9])?)*$"
