+++ tests/system/providers/cncf/kubernetes/example_kubernetes.py
@@ -28,7 +28,7 @@ from kubernetes.client import models as k8s
 from airflow import DAG
 from airflow.kubernetes.secret import Secret
 from airflow.operators.bash import BashOperator
-from airflow.providers.cncf.kubernetes.operators.kubernetes_pod import KubernetesPodOperator
+from airflow.providers.cncf.kubernetes.operators.pod import KubernetesPodOperator
 
 # [START howto_operator_k8s_cluster_resources]
 secret_file = Secret("volume", "/etc/sql_conn", "airflow-secrets", "sql_alchemy_conn")
